Update models and controllers for tblPY1, tblcliente, tblarticulo, tblcompra, tblcolocacion, and tblfactura
This commit adds and modifies models and controllers for the following tables:
- tblPY1
- tblcliente
- tblarticulo
- tblcompra
- tblcolocacion
- tblfactura

The models and controllers are necessary for managing data related to users, clients, articles, purchases, placements, and invoices. The changes include adding new models and controllers, as well as updating existing ones to improve functionality and adhere to specific requirements.

Note: This commit assumes that the changes are being made as part of the ongoing development of a Laravel application.